Added logging by SLF4J for 'Client'

+++ src/main/java/ua/aleks4ay/kiyv/copiller_db/reader/CopyClient.java
@@ -1,5 +1,7 @@
 package ua.aleks4ay.kiyv.copiller_db.reader;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import ua.aleks4ay.kiyv.domain.dao.ClientDao;
 import ua.aleks4ay.kiyv.domain.dao.ClientDaoJdbc;
 import ua.aleks4ay.kiyv.domain.dbf.ClientDbf;
@@ -11,11 +13,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import static ua.aleks4ay.kiyv.log.ClassNameUtil.getCurrentClassName;
+
 public class CopyClient {
 
     private static final CopyClient copyClient = new CopyClient();
     private static final ClientDao clientDao = new ClientDaoJdbc();
     private static final ClientDbf clientDbfReader = new ClientDbfReader();
+    private static final Logger log = LoggerFactory.getLogger(getCurrentClassName());
 
     private CopyClient() {
     }
@@ -27,6 +32,7 @@ public class CopyClient {
     public static void main(String[] args) {
         long start = System.currentTimeMillis();
 
+        log.debug("Executing doCopyClient.");
         CopyClient.getInstance().doCopyNewRecord();
 
         long end = System.currentTimeMillis();
@@ -34,23 +40,22 @@ public class CopyClient {
     }
 
     public void doCopyNewRecord() {
-//        ClientDao clientDao = new ClientDaoJdbc();
-//        ClientDbf clientDbfReader = new ClientDbfReader();
-
+        log.debug("Start writing 'C L I E N T S'.");
+        log.debug("Check availability write to DataBase.");
         while ( clientDao.isBlocking()) {
             try {
+                log.debug("Wait 5 seconds.");
                 Thread.sleep(5000);
             } catch (InterruptedException e) {
-                e.printStackTrace();
+                log.warn("Exception during try 'Thread.sleep' 5 seconds.", e);
             }
         }
-        clientDao.setBlock();
-
-        System.out.println("start write  'C L I E N T S'");
+//        log.debug("Set lock to DataBase for another threads.");
+        clientDao.setLock();
 
         List<Client> listNewClient = new ArrayList<>();
-        List<Client> listDeletingClient = new ArrayList<>();
         List<Client> listUpdatingClient = new ArrayList<>();
+
         Map<String, String> oldClient = clientDao.getAll()
                 .stream()
                 .collect(Collectors.toMap(Client::getId, Client::getClientName));
@@ -69,21 +74,22 @@ public class CopyClient {
             }
         }
 
-        clientDao.saveAll(listNewClient);
-        clientDao.updateAll(listUpdatingClient);
-        clientDao.deleteAll(oldClient.keySet());
-
-        System.out.println("Added " + listNewClient.size() +
-                ", Removed " + oldClient.size() +
-                ", Updated " + listUpdatingClient.size() + " clients.");
-        System.out.println("End write to DB\n");
-        try {
-            clientDao.setUnblock();
-
-            Thread.sleep(500);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
+        if (listNewClient.size() > 0) {
+            log.debug("Save to DataBase. Must be added {} new clients.", listNewClient.size());
+            clientDao.saveAll(listNewClient);
+        }
+        if (listUpdatingClient.size() > 0) {
+            log.debug("Write change to DataBase. Must be updated {} clients.", listUpdatingClient.size());
+            clientDao.updateAll(listUpdatingClient);
+        }
+        if (oldClient.size() > 0) {
+            log.debug("Delete old client from DataBase. Must be deleted {} clients.", oldClient.size());
+            clientDao.deleteAll(oldClient.keySet());
         }
 
+//        log.debug("Set unlock to DataBase.");
+        clientDao.setUnlock();
+
+        log.debug("End writing 'C L I E N T S'.");
     }
 }
